Organize files structure
